Deleted English/Control.pm
English/Control.pm was just an idea to use $^{ARG} instead of $ARG for
punctuation variables, but it wasn't used, and I never
kept up with changes to it in perl's English.pm. If Perl had started with
this it might have been better than the combination of punctuation
variables and the aliases from English.pm, but that was a long time ago.