Moved load_tables to FileMaker_ODBC object
Previously, the load_tables method of
Actium::O::Files::FileMaker_ODBC.pm was in its own procedural
module Actium::Files::FileMaker_ODBC. With other changes
it was now easy to move this to be a method, reducing the need
for a duplicate file. This added load_tables to the class and
changed the invocation for all the modules that call it.