Changed cache folder location & modified mainhelp
Changed default cache folder location in Config::Signup
to ~/.actium in the user's home directory. This will override
the default in ::SQLite.

In ::SQLite, added better checks for when the directory doesn't
exist or when it's not really a directory.

Changed _mainhelp in ::Cmd to indent the list by two spaces.