Used own cache instead of Flyweight role
To ensure that there was only one cache and it was used regularly,
removed use of the MooseX::Role::Flyweight. Now it uses its own
caches, one for from_str for time strings, and one for from_num for
time numbers.

Added named time numbers (e.g., "NOON_YESTERDAY") which can be used
in place of time strings.

Added new apbx marker "z" for noon tomorrow (otherwise it was an anomaly,
a legal time -- used to have a "largest possible time" --
but one without a marker).