Allowing different library for Actium::Cmd
Actually this just moves the "Actium::Cmd" part into a constant,
but this would make it easier to change this later if desired.
The idea is that I may want one library for real commands that
are permanent, and another for quick and dirty scripts that rely
on the libraries.