No longer merge on BUILD in TripCollection
The reason is that this is premature -- we don't want to prematurely merge
trips that end up being across schedule-days.  (For example, if on weekends we
have trips running every 20 minutes and on weekends they run every 15, then
once an hour they might be identical. We don't want to prematurely merge the
trips so they end up being the same across the days when there's no good reason
to do so.)

Split and re-sort @found_days so we don't end up with days like '12453' when it
combines 1245 and 3

Generally improve across-day merging and dividing, making sure that trips are
cloned and not the same across days.

Remove old commented-out merging code