Add specialdays; change caches to hashrefs
Changed the caches in some field to state hashrefs instead of state hashes.
Turns out this isn't necessary but it works, so whatever.
Those should probably be made lazy/build attributes anyway.

Added a predicate to as_shortcode.

Fixed abbreviations. Added as_transitinfo for tabskeds.

Added is_equal_to, which relies on the objects being the same.
In theory this should always be the case but storage may mean it won't,
so this should probably be tested.