Removed BUILDARGS in O::Dir
So MooseX::Storage didn't know about the positional arguments in
BUILDARGS, and that kept breaking. So I removed it, since new should
only be invoked from here anyway due to its flyweight-ness.

There are other problems with that (it shouldn't be invoking new...)
but it simplifies the code so I'm leaving it.