Fix errors producing schedules
Elimiated the extra lines to height, because height is used
for calculating which rows to output. Will still need to
adjust heights when notes are used with overlong timetables
(ones that are too long and have to span frames).

Made an attempt at creating responsive HTML tables but it just isn't working...