Supplemental Calendar note changes
Change the threshold for "every Wednesday except..." so that
it can be off 1/3 of the time before it starts using individual dates,
instead of 1/4 of the time.

Changed it so that if there are more than two characters in the days,
it uses the generic "A" for the note rather than the days.
This avoids overly long notes like WTh-5 and MWF-1.