SuppCalendar: change formula for display dates
Previously, it had a bias towards displaying "on" dates, rather than
"Every Wednesday except...".
It would display all the "on" dates if there were more than 1/3 off.
This is tedious with the long lists of on dates with the regular
schedules, so now it uses whichever has fewer dates, ons or offs.