::Storage::File - methods that throw exceptions
Add a bunch of methods that throw exceptions on failure, but
otherwise just do what Path::Class does.

 - remove (this had been "delete" but now it replaces Path::Class's remove)
 - touch
 - open
 - copy_to
 - move_to

Also, put in stub openr, openw, opena methods that don't do anything but just
croak, because they don't allow for an encoding and it seems like a bad idea to
allow defaulting to nothing.

Added tests for the new methods.