Set primary attributes of each column
Previously, the primary attribute of each column was set separately --
the most frequent lines, the most frequent destination, etc. was picked.
That led to a situation where for example the most frequent line had
multiple destinations, so a different line had the most frequent
destination -- and every single time had a footnote. Now, it chooses the
most common *combination* as the primary, so that the number of
footnotes is reduced.