Fix headers of "continued" timetables
Lots of timetables have had problems where the header overflowed on
tables where "continued" is needed. The actual problem is not leaving
enough room for the header, but that's a hard problem to fix. For now
leaving space at the bottom of timetables with "continued" solves the
issue while in practice having no negative impact (if there were a very
long schedule where it *almost* needed three+ columns, but not quite,
then it might use an extra page. But there are none like that right now)