Add a clone of the Actium FileMaker database
This is a copy of the FileMaker database saved as a "clone" from within
FileMaker, which means it has the layouts, scripts, database structure,
calculated fields, etc., but no actual data in it. Data will be uploaded
as CSV files separately (on the ground that probably nobody will need
the FileMaker file actually).